Enum bitgrin_wallet_controller::ErrorKind

pub enum ErrorKind {
    LibTX(ErrorKind),
    Impls(ErrorKind),
    LibWallet(ErrorKind, String),
    Keychain(Error),
    Transaction(Error),
    Secp,
    FileWallet(&'static str),
    IO,
    Format,
    Node(ErrorKind),
    Hyper,
    Uri,
    DuplicateTransactionId,
    WalletSeedExists(String),
    WalletSeedDoesntExist,
    Encryption,
    Mnemonic,
    ArgumentError(String),
    GenericError(String),
}

Wallet errors, mostly wrappers around underlying crypto or I/O errors.

Variants
LibTX(ErrorKind)

LibTX Error

Impls(ErrorKind)

Impls error

LibWallet(ErrorKind, String)

LibWallet Error

Keychain(Error)

Keychain error

Transaction(Error)

Transaction Error

Secp

Secp Error

FileWallet(&'static str)

Filewallet error

IO

Error when formatting json

Format

Error when formatting json

Node(ErrorKind)

Error when contacting a node through its API

Hyper

Error originating from hyper.

Uri

Error originating from hyper uri parsing.

DuplicateTransactionId

Attempt to use duplicate transaction id in separate transactions

WalletSeedExists(String)

Wallet seed already exists

WalletSeedDoesntExist

Wallet seed doesn't exist

Encryption

Enc/Decryption Error

Mnemonic

BIP 39 word list

ArgumentError(String)

Command line argument error

GenericError(String)

Other